Might as well remove unused hexgrid classes
