--- src/common/components/hexgrid/index.js
@@ -1,6 +1,0 @@
-import Hex from './Hex';
-import HexGrid from './HexGrid';
-import Layout from './Layout';
-import HexUtils from './HexUtils';
-
-export { HexGrid, Hex, Layout, HexUtils };
